moved feeds.json file to '.config/core-rss', and added 'open in browser' feature
open in browser is useful when the feed doesn't provide the full content
to read, so we can read the full stuff in the browser.

it works by getting the id of the current selected item in any feed,
then using a native os command to open with the default app.

i also added 2 extra commands (commented of course) for windows and
macos, but i don't use windows nor macos, so i cant test for now